quick rejigger to make __init__.py much simpler
and to not lose github history
